+++ resources/views/components/payfast-jetstream-subscriptions.blade.php
@@ -1,35 +1,147 @@
-<x-jet-form-section submit="updateSubscriptionInformation">
+<x-jet-action-section>
     <x-slot name="title">
         {{ __('Subscription Information') }}
     </x-slot>
 
     <x-slot name="description">
-        {{ __('Update your subscription and credit card information.') }}
+        {{ __('View/Update subscription information.') }}
     </x-slot>
 
-    <x-slot name="form">        
-        <!-- Credit Card Token -->
-        <div class="col-span-6 sm:col-span-4">
-            <x-jet-label for="credit_card_token" value="{{ __('Credit Card Token') }}" />
-            <x-jet-input id="credit_card_token" type="text" class="mt-1 block w-full" wire:model.defer="state.credit_card_token" autocomplete="credit_card_token" />
-            <x-jet-input-error for="credit_card_token" class="mt-2" />
-
-            {!! $updateCard !!}
-
-            <div wire:ignore>
-            {!! $createSubscription !!}
-            </div>
+    <x-slot name="content">
+        <div class="max-w-xl text-sm text-gray-600">
             
-        </div>        
-    </x-slot>
+            @if ($user->onGenericTrial() && $user->subscriptions()->active()->count() == 0)
+                <h3 class="text-lg font-medium text-gray-900">
+                    You are current on trial till the {{ $user->trialEndsAt()->format('jS F Y') }}
+                </h3>
+                <div class="mt-3 max-w-xl text-sm text-gray-600">
+                    <p>
+                        If you subscribe now the first payment is due on the {{ $user->trialEndsAt()->addDay()->format('jS F Y'); }}.
+                    </p>
+                </div>
+            @endif
 
-    <x-slot name="actions">
-        <x-jet-action-message class="mr-3" on="saved">
-            {{ __('Saved.') }}
-        </x-jet-action-message>
+            @if ($user->subscriptions()->active()->count() > 0)
+                <h3 class="text-lg font-medium text-gray-900">
+                    You are subscribed to the
+                    {{ $user->subscriptions()->active()->first()->name }} plan.
+                </h3>
+                <div class="mt-3 max-w-xl text-sm text-gray-600">
+                    <p>
+                        The next payment is due
+                        {{ $user->subscriptions()->active()->first()->next_bill_at->format('Y-m-d') }}.
+                    </p>
+                </div>
+            @endif
+            
+            @if (!$user->onGenericTrial() && $user->subscriptions()->active()->count() == 0)
+                <h3 class="text-lg font-medium text-gray-900">
+                    You are not currently subscribed to any plan.
+                </h3>
+                <div class="mt-3 max-w-xl text-sm text-gray-600">
+                    <p>
+                        {{ __('Please select from our list of plans below') }}
+                    </p>
+                </div>
+            @endif
+        </div>
 
-        <x-jet-button wire:loading.attr="disabled" wire:target="credit_card_token">
-            {{ __('Save') }}
-        </x-jet-button>
+        <!-- Subscription Action Buttons -->
+        <div class="mt-5">            
+            @if ($user->subscriptions()->active()->count() > 0)
+                <x-jet-secondary-button
+                    style="color: red;"
+                    wire:click="confirmCancelSubscription"
+                    wire:loading.attr="disabled">
+                    {{ __('Cancel') }}
+                </x-jet-secondary-button>
+
+                <x-jet-secondary-button
+                    style="color: blue;"
+                    >
+                    {{ __('Update Card') }}
+                </x-jet-secondary-button>                                                
+            @else
+                <div class="flex">
+                    <select
+                        wire:model="plan"
+                        name="plan"
+                        class="mt-1 block pl-3 pr-10 py-2 text-base border-gray-300 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm rounded-md"                        
+                        >
+                        @foreach (config('payfast.plans') as $key => $value)
+                            <option value="{{ $key }}">{{ $value['name'] }}</option>
+                        @endforeach
+                    </select>
+
+                    <x-jet-secondary-button
+                        class="ml-2 align-middle h-9 mt-2"
+                        style="color: green;"
+                        wire:click="displayCreateSubscription"
+                        >
+                        {{ __('Subscribe') }}
+                    </x-jet-secondary-button>
+
+                    <div wire:loading
+                        class="ml-2 align-middle mt-3"
+                    >
+                        Please wait...
+                    </div>
+
+                </div>
+            @endif
+
+            
+        </div>
+        <!-- End Subscription Action Buttons -->
+
+        <!-- Create Subscription Modal -->
+        @if($displayingCreateSubscription)        
+            <script>
+                window.payfast_do_onsite_payment({"uuid":"{{ $identifier }}"})                
+                window.addEventListener("message", refreshComponent);
+            </script>
+        @endif  
+
+        @push('body')
+            <script>        
+            const refreshComponent = () => {
+                        console.log('Refreshing subscription status')
+                        Livewire.emit('billingUpdated')
+                    }
+            </script>
+        @endpush
+
+        <!-- Cancel Subscription Confirmation Modal -->
+        <x-jet-dialog-modal wire:model="confirmingCancelSubscription">
+
+            <x-slot name="title">
+                {{ __('Cancel Subscription') }}
+            </x-slot>
+
+            <x-slot name="content">
+                {{ __('Are you sure you want to cancel your Best Agent subscription? Once your subscription is cancelled, your Best Agent profile, rankings, and property listings will be hidden.') }}
+                <br><br>
+                {{ __('The last date until which your rankings and sold properties will show is the ') }}                
+            </x-slot>
+
+            <x-slot name="footer">
+                <div wire:loading
+                        class="mr-2 align-middle mt-3"
+                    >
+                        Please wait...
+                </div>
+
+                <x-jet-secondary-button wire:click="$toggle('confirmingCancelSubscription')" wire:loading.attr="disabled">
+                    {{ __('Cancel') }}
+                </x-jet-secondary-button>
+
+                <x-jet-danger-button class="ml-2" wire:click="cancelSubscription" wire:loading.attr="disabled">
+                    {{ __('Cancel Subscription') }}
+                </x-jet-danger-button>
+            </x-slot>
+
+        </x-jet-dialog-modal>
+        <!-- Cancel Subscription Confirmation Modal -->
+                
     </x-slot>
-</x-jet-form-section>
+</x-jet-action-section>